refactored service packge to usecase package, deleted cooking method classes
