+++ src/main/java/ChouTuanEat/controller/LoginController.java
@@ -1,7 +1,7 @@
 package ChouTuanEat.controller;
 
 import ChouTuanEat.entity.User;
-import ChouTuanEat.service.UserService;
+import ChouTuanEat.usecase.UserService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
