--- src/main/java/ChouTuanEat/entity/CookMethods.java
@@ -1,5 +1,0 @@
-package ChouTuanEat.entity;
-
-public enum CookMethods {
-    BOILED, STEAMED, GRILLED, FRIED
-}
