+++ src/main/java/ChouTuanEat/usecase/Impl/IngredientsServicelmpl.java
@@ -1,9 +1,8 @@
-package ChouTuanEat.service.Impl;
+package ChouTuanEat.usecase.Impl;
 
-import ChouTuanEat.entity.Dishes;
 import ChouTuanEat.entity.Ingredients;
 import ChouTuanEat.repository.IngredientsRepository;
-import ChouTuanEat.service.IngredientsService;
+import ChouTuanEat.usecase.IngredientsService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
