+++ src/main/java/ChouTuanEat/usecase/Impl/UserServiceImpl.java
@@ -1,8 +1,8 @@
-package ChouTuanEat.service.Impl;
+package ChouTuanEat.usecase.Impl;
 
         import ChouTuanEat.entity.User;
         import ChouTuanEat.repository.UserRepository;
-        import ChouTuanEat.service.UserService;
+        import ChouTuanEat.usecase.UserService;
         import org.springframework.beans.factory.annotation.Autowired;
         import org.springframework.stereotype.Service;
