+++ src/main/java/ChouTuanEat/usecase/IngredientsService.java
@@ -1,4 +1,4 @@
-package ChouTuanEat.service;
+package ChouTuanEat.usecase;
 
 import ChouTuanEat.entity.Ingredients;
