+++ src/test/java/ChouTuanEat/controller/ViewFavouriteDishControllerTest.java
@@ -2,13 +2,11 @@ package ChouTuanEat.controller;
 
 import ChouTuanEat.entity.Dishes;
 import ChouTuanEat.entity.favourites;
-import ChouTuanEat.service.favouriteService;
-import com.fasterxml.jackson.core.JsonProcessingException;
+import ChouTuanEat.usecase.favouriteService;
 import com.fasterxml.jackson.core.type.TypeReference;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import org.junit.jupiter.api.Test;
 import org.junit.runner.RunWith;
-import org.slf4j.helpers.MessageFormatter;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
 import org.springframework.boot.test.context.SpringBootTest;
@@ -16,11 +14,8 @@ import org.springframework.http.MediaType;
 import org.springframework.test.context.junit4.SpringRunner;
 import org.springframework.test.web.servlet.MockMvc;
 import org.springframework.test.web.servlet.MvcResult;
-import org.springframework.test.web.servlet.ResultMatcher;
 import org.springframework.util.StringUtils;
 
-import java.io.UnsupportedEncodingException;
-import java.text.MessageFormat;
 import java.util.List;
 
 import static org.junit.jupiter.api.Assertions.*;
