+++ src/test/java/ChouTuanEat/service/Impl/favouriteServiceImplTest.java
@@ -4,7 +4,9 @@ package ChouTuanEat.service.Impl;
 import ChouTuanEat.entity.Dishes;
 import ChouTuanEat.entity.favourites;
 import ChouTuanEat.repository.FavouriteRepository;
-import ChouTuanEat.service.DishesService;
+import ChouTuanEat.usecase.DishesService;
+import ChouTuanEat.usecase.Impl.DishesServicelmpl;
+import ChouTuanEat.usecase.Impl.favouriteServiceImpl;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
